Fixing

+++ LibraryWebApi/Library.Application/UseCases/BookUseCases/DeleteBookUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Library.Application.Exceptions;
 using Library.Domain.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
@@ -6,18 +7,22 @@ namespace Library.Application.UseCases.BookUseCases
 {
     public class DeleteBookUseCase
     {
-        private readonly IMapper _mapper;
         private readonly IUnitOfWork _unitOfWork;
 
-        public DeleteBookUseCase(IMapper mapper,
-            IUnitOfWork unitOfWork)
+        public DeleteBookUseCase(IUnitOfWork unitOfWork)
         {
-            _mapper = mapper;
             _unitOfWork = unitOfWork;
         }
 
         public async Task DeleteBook([FromRoute] int id)
         {
+            var existingBook = await _unitOfWork.Book.GetByIdAsync(id);
+
+            if (existingBook is null)
+            {
+                throw new EntityNotFoundException($"{existingBook} is not found in database.");
+            }
+
             await _unitOfWork.Book.DeleteAsync(id);
 
             await _unitOfWork.SaveChangesAsync();
